added drag n drop code to tank. each tank moves independently
