+++ scenes/Main.tscn
@@ -1,0 +1,23 @@
+[gd_scene load_steps=4 format=2]
+
+[ext_resource path="res://scenes/Main.cs" type="Script" id=1]
+[ext_resource path="res://scenes/map1.tscn" type="PackedScene" id=2]
+[ext_resource path="res://scenes/Player.tscn" type="PackedScene" id=3]
+
+[node name="Main" type="Node"]
+script = ExtResource( 1 )
+
+[node name="map1" parent="." instance=ExtResource( 2 )]
+
+[node name="playerUnits" type="Node" parent="."]
+
+[node name="player1" parent="playerUnits" groups=[
+"playerUnits",
+] instance=ExtResource( 3 )]
+position = Vector2( 32, 96 )
+
+[node name="player2" parent="playerUnits" groups=[
+"playerUnits",
+] instance=ExtResource( 3 )]
+position = Vector2( 160, 224 )
+
