+++ scenes/map1.tscn
@@ -1,9 +1,8 @@
-[gd_scene load_steps=6 format=2]
+[gd_scene load_steps=5 format=2]
 
-[ext_resource path="res://scenes/map.cs" type="Script" id=1]
+[ext_resource path="res://scenes/map1.cs" type="Script" id=1]
 [ext_resource path="res://assets/terrain.tres" type="TileSet" id=2]
 [ext_resource path="res://assets/test.png" type="Texture" id=3]
-[ext_resource path="res://scenes/Player.tscn" type="PackedScene" id=4]
 
 [sub_resource type="TileSet" id=1]
 0/name = "Grid"
@@ -17,7 +16,7 @@
 0/shapes = [  ]
 0/z_index = 0
 
-[node name="Node2D" type="Node2D"]
+[node name="map1" type="Node2D"]
 script = ExtResource( 1 )
 
 [node name="terrain" type="TileMap" parent="."]
@@ -29,15 +28,4 @@ tile_data = PoolIntArray( 0, 1, 0, 1, 0, 0, 2, 1, 0, 3, 0, 0, 4, 1, 0, 65536, 10
 tile_set = SubResource( 1 )
 format = 1
 tile_data = PoolIntArray( 0, 0, 0, 1, 0, 0, 2, 0, 0, 3, 0, 0, 4, 0, 0, 65536, 0, 0, 65537, 0, 0, 65538, 0, 0, 65539, 0, 0, 65540, 0, 0, 131072, 0, 0, 131073, 0, 0, 131074, 0, 0, 131075, 0, 0, 131076, 0, 0, 196608, 0, 0, 196609, 0, 0, 196610, 0, 0, 196611, 0, 0, 196612, 0, 0, 262144, 0, 0, 262145, 0, 0, 262146, 0, 0, 262147, 0, 0, 262148, 0, 0 )
-
-[node name="Player" type="Node" parent="."]
-
-[node name="Tank1" parent="Player" groups=[
-"Player",
-] instance=ExtResource( 4 )]
-position = Vector2( 160, 160 )
-
-[node name="Tank2" parent="Player" groups=[
-"Player",
-] instance=ExtResource( 4 )]
